refactor(frontend): normalize telemetry names
