+++ frontend/app/.server/utils/telemetry-utils.ts
