+++ frontend/nodemon.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json.schemastore.org/nodemon",
-  "exec": "tsx --env-file-if-exists=.env --import ./app/.server/opentelemetry.ts ./app/.server/express/server.ts",
+  "exec": "tsx --env-file-if-exists=.env --import ./app/.server/telemetry.ts ./app/.server/express/server.ts",
   "ext": "ts",
   "watch": [".env", "./app/.server/express/**/*.ts"]
 }
