+++ frontend/tests/utils/telemetry-utils.test.ts
@@ -3,7 +3,7 @@ import { SpanStatusCode, trace } from '@opentelemetry/api';
 import { describe, expect, it, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
-import * as InstrumentationUtils from '~/.server/utils/instrumentation-utils';
+import { handleSpanException, withSpan } from '~/.server/utils/telemetry-utils';
 import { AppError } from '~/errors/app-error';
 import { ErrorCodes } from '~/errors/error-codes';
 
@@ -15,7 +15,7 @@ describe('InstrumentationUtils', () => {
       const mockSpan = mock<Span>();
       const error = new AppError('test error message', ErrorCodes.TEST_ERROR_CODE);
 
-      InstrumentationUtils.handleSpanException(error, mockSpan);
+      handleSpanException(error, mockSpan);
 
       expect(mockSpan.recordException).toHaveBeenCalledWith({
         name: error.name,
@@ -34,7 +34,7 @@ describe('InstrumentationUtils', () => {
       const mockSpan = mock<Span>();
       const error = new Error('regular error');
 
-      InstrumentationUtils.handleSpanException(error, mockSpan);
+      handleSpanException(error, mockSpan);
 
       expect(mockSpan.recordException).toHaveBeenCalledWith({
         name: error.name,
@@ -53,7 +53,7 @@ describe('InstrumentationUtils', () => {
       const mockSpan = mock<Span>();
       const error = new Response(null, { status: 500 });
 
-      InstrumentationUtils.handleSpanException(error, mockSpan);
+      handleSpanException(error, mockSpan);
 
       expect(mockSpan.recordException).not.toHaveBeenCalled();
       expect(mockSpan.setStatus).not.toHaveBeenCalled();
@@ -63,7 +63,7 @@ describe('InstrumentationUtils', () => {
       const mockSpan = mock<Span>();
       const error = new AppError('Something went wrong');
 
-      const returnedError = InstrumentationUtils.handleSpanException(error, mockSpan);
+      const returnedError = handleSpanException(error, mockSpan);
 
       expect(returnedError).toEqual(error);
     });
@@ -75,7 +75,7 @@ describe('InstrumentationUtils', () => {
       const mockTracer = mock<Tracer>({ startSpan: vi.fn().mockReturnValue(mockSpan) });
       vi.mocked(trace.getTracer).mockReturnValue(mockTracer);
 
-      await InstrumentationUtils.withSpan('test span', () => {});
+      await withSpan('test span', () => {});
 
       expect(mockTracer.startSpan).toHaveBeenCalledWith('test span');
       expect(mockSpan.end).toHaveBeenCalled();
@@ -87,7 +87,7 @@ describe('InstrumentationUtils', () => {
       vi.mocked(trace.getTracer).mockReturnValue(mockTracer);
       const mockFn = vi.fn().mockResolvedValue('test');
 
-      const result = await InstrumentationUtils.withSpan('test span', mockFn);
+      const result = await withSpan('test span', mockFn);
 
       expect(mockFn).toHaveBeenCalledWith(mockSpan);
       expect(result).toEqual('test');
@@ -105,7 +105,7 @@ describe('InstrumentationUtils', () => {
       const error = new Error('test error');
       const mockFn = vi.fn().mockRejectedValue(error);
 
-      await expect(InstrumentationUtils.withSpan('test span', mockFn)).rejects.toThrow(error);
+      await expect(withSpan('test span', mockFn)).rejects.toThrow(error);
 
       expect(mockSpan.recordException).toHaveBeenCalledWith({
         name: error.name,
